unsquashfs: Add a -t[est] option, which only tests the image
Taken from this patch by Zdenek Behan:
http://sourceforge.net/p/squashfs/patches/15/

Introduce a -t[est] option for unsquashfs utility. This prevents creation of
anything, anywhere, and redirects data writes to /dev/null, and returns an
error code if something goes wrong, or 0 if the filesystem uncompresses fine,
acting like fsck. It's similar approach to what is used in cramfsck.

While this method may be a bit hacky, it's generic enough to span all fs
versions and is probably easilly backportable.

I'm not the author of it, but I don't have the original author's mail
to give proper credit.